+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## 1.2.30
 - Disable "Block Useless Perks" in IQ Addons by default for new and existing users (if you turn it back on it will stay enabled)
+- Also disable Remove Reloading Screen's missing texture substitution on 26.2+ Vulkan to stop more crashes